Make delegate and data source visible in Interface Builder
Although UIScrollView .delegate etc. are visible in Interface Builder without the IBOutlet attribute, custom properties still need it, thus in order for the empty state to be completely configurable in Interface Builder the two properties need the attributes. 

This will become especially handy when separating the empty state handling into a separate NSObject subclass that can be reused between controllers, in which case the latter does not want/need a direct connection to the UIScrollView it is providing the empty state for, or it might even provide the empty state for multiple at a time.