Set accessibility identifiers instead of accessibility labels
Accessibility labels should be localized and since those views (image view, title label, detail label and button) are not publicly exposed it’s impossible to change them.

Also, the best accessibility label for a UILabel is its text, which is the default when you don’t set any `accessibilityLabel` property.