Fixes gestures not passing through UITableViewWrapperView on iOS 7
This fixes #47

In a nutshell, this pull request toggles `UITableViewWrapperView`'s
`userInteractionEnabled` property off or on when the `emptyDataSetView`
appears or disappears.

The main concern here I suppose is the check for
"UITableViewWrapperView" since that is a private API. This call is
guarded by a version check, so shouldn't break in future releases, but
I'm not sure if checking for this class could cause an app to be
rejected.

For what it's worth, validating an Archive containing this code didn't
cause any private API flags to be raised.