Revert #259
If a custom view has a UIControl subclass, the -gestureRecognizer:shouldReceiveTouch: call is necessary to allow touches to pass through to the control immediately. Without this call, touches are delayed, and the touch up action on the control doesn't fire unless the control is tapped for an extended period of time.

The call here is checking to see if the tap gesture on the DZNEmptyDataSetView should receive touches. If the tapped view is a UIControl, the gesture recognizer shouldn't receive any touches and should instead pass them along to the UIControl for processing.